OMNIDRUG-AI/backend/requirements.txt
# Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
drf-spectacular==0.27.0

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0
redis==5.0.1

# Celery
celery==5.3.4
flower==2.0.1

# Scientific
rdkit==2023.9.4
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4
biopython==1.82

# Authentication
djangorestframework-simplejwt==5.3.1

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.2.0
